// src/enums/userPermissions.enum.ts
export enum UserPermissionsEnum {
    SASP_MANAGE_SETTINGS = "sasp_manage_settings",
    SASP_MANAGE_SASP_USERS = "sasp_manage_sasp_users",
    SASP_APPROVE_ORGANIZATIONS = "sasp_approve_organizations",
    SASP_VIEW_ALL_ORGANIZATIONS = "sasp_view_all_organizations",
    SASP_VIEW_ALL_DONEE_ACCOUNTS = "sasp_view_all_donee_accounts",
    SASP_VIEW_ALL_USERS = "sasp_view_all_users",
    SASP_VIEW_ALL_REQUESTS = "sasp_view_all_requests",
    SASP_MANAGE_ALL_REQUESTS = "sasp_manage_all_requests",
    SASP_GENERATE_REQUEST_LOAR = "sasp_generate_request_loar",
    SASP_GENERATE_REQUEST_INVOICE = "sasp_generate_request_invoice",
    VIEW_ORGANIZATION_REQUESTS = "view_organization_requests",
    MANAGE_ORGANIZATION_DONEE_ACCOUNT = "manage_organization_donee_account",
    VIEW_ORGANIZATION_INFO = "view_organization_info",
    MANAGE_ORGANIZATION_INFO = "manage_organization_info",
    MANAGE_ORGANIZATION_USERS = "manage_organization_users",
    MANAGE_DONEE_ACCOUNT = "manage_donee_account",
    MANAGE_DONEE_ACCOUNT_USERS = "manage_donee_account_users",
    MANAGE_REQUESTS = "manage_requests",
    ATTACH_FILES_TO_REQUESTS = "attach_files_to_requests",
}